requirements.txt
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
coveralls==1.3.0
Django==2.0.7
djangorestframework==3.8.2
docopt==0.6.2
idna==2.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
virtualenv==16.0.0
